refactor: rename telainicial to inicio
